Database schema and some addtional refactoring

+++ PersonalAccounting.Domain/Entities/BaseEntity.cs
@@ -3,11 +3,11 @@
 namespace PersonalAccounting.Domain.Entities
 {
     public abstract class BaseEntity
-    {
-        public int Id { get; set; }
+    {        
+        public long Id { get; set; }
         public DateTime CreatedAt { get; set; }
-        public int CreatedBy { get; set; }
+        public long CreatedBy { get; set; }
         public DateTime UpdatedAt { get; set; }
-        public int UpdatedBy { get; set; }
+        public long UpdatedBy { get; set; }
     }
 }
